requirements.txt
sqlalchemy
psycopg2-binary
pymysql
snowflake-connector-python
python-dotenv
pydantic

crewai[google-genai]
langchain
chromadb
sentence-transformers
tf-keras

google-genai